requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
openai>=1.40.0
graphviz>=0.20.3
jinja2>=3.1.4
huggingface_hub>=0.32.0
Pillow>=10.4.0
numpy>=1.26.4
imageio>=2.36.0
imageio-ffmpeg>=0.5.1
